requirements.txt
redis
redis-py
neo4j-driver
neobolt
rdkit
pandas